refactor(cli): весь конфиг вебпака перенёс в папку webpack
