+++ packages/cli/config/entry/entry.ts
