+++ packages/cli/config/entry/fake-entry.ts
