+++ packages/cli/config/entry/index.ts
@@ -1,0 +1,1 @@
+export { getEntryPoints } from './entry';
