+++ packages/cli/config/plugins/create-federation-plugin.ts
@@ -2,7 +2,7 @@ import { NodeFederationPlugin } from '@module-federation/node';
 import { container } from 'webpack';
 
 import { getFederationPluginOptions } from './get-federation-plugin-options';
-import type { Settings } from '../types';
+import type { Settings } from '../../types';
 
 /** Создаёт и возвращает экземпляр ModuleFederationPlugin */
 export function createFederationPlugin(settings: Settings): NodeFederationPlugin {
