+++ packages/cli/config/plugins/index.ts
@@ -1,0 +1,1 @@
+export { getPlugins } from './get-plugins';
